refactor: reallocate functions and modify alias command
